requirements.txt
appdirs==1.4.4
boto3==1.24.28
botocore==1.27.59
brotlipy==0.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
idna==3.4
importlib-metadata==6.6.0
jmespath==0.10.0
packaging==23.1
pip==23.0.1
pycparser==2.21
pyee==8.1.0
pyopenssl==23.0.0
pyppeteer==1.0.2
pysocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.6.0
selenium==3.141.0
setuptools==67.8.0
six==1.16.0
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.15
webdriver-manager==3.8.6
websockets==10.4
wheel==0.38.4
win_inet_pton==1.1.0
zipp==3.15.0